Implement LeadText
